// src/types/api.d.ts
export type Battery = {
  'health': '良好',
  'level': '100%',
  'plugged': 'AC',
  'scale': '100%',
  'status': '充电中',
  'temperature': '26.70℃',
  'voltage': '4.47V'
}

export type Sms = {
  'content': string,
  'date': number,
  'name': '未知号码',
  'number': string,
  'sim_id': number,
  'sub_id': number,
  'type': number,
  'typeImageId': number
}